Add timing event and verification handlers to raceweb app with structured logging
